Break commands out of ctl.js
